+++ sark/exceptions/base.py
