+++ sark/exceptions/function.py
@@ -1,4 +1,4 @@
-from base_exceptions import *
+from base import *
 
 class SarkFunctionException(SarkError):
     pass
