+++ sark/exceptions/switch.py
@@ -1,4 +1,4 @@
-from base_exceptions import *
+from base import *
 
 class SarkSwitchError(SarkError):
     pass
